snowflake_managed_mcp/requirements.txt
alembic
snowflake-sqlalchemy
pandas